Differentiated between db unique name and db name
Until now, the ORA_DBNAME variable was set to the value passed to the script. But in case of databases where the unique name differs from the db name, this meant that the ORA_DBNAME was set to the database unique name.
This change, introduces the ORACLE_UNQNAME and stores the unique name in the variable, while the ORA_DBNAME is set to the database name.

This change was also required for the db cloning scripts used at Base as these require access to the database name to perform the rman duplicate.

The rac_aliasses where modified to change the use of the ORA_DBNAME variable to the ORACLE_UNQNAME variable