When the db_name is not set in the cluster config it was left blank instead of setting it to the db unique name value in the cluster config
When in the GI the db unique name does not differ from the db name, the -n <db name> parameter is not mandatory and when not specified it is left blank (and not defaulted to the db unique name).
As result the ORA_DBNAME was not set by godb.
This fix changes this and sets in such case the ORA_DBNAME to the ORACLE_UNQNAME