modified gogrid to be able to work in environments where asm is not used
Before this change, the script only grepped on ASM in the /etc/oratab file and used the oraenv script to set the environment.
Now, the script will still grep for ASM in the /etc/oratab file, but when no entry found it will set the required variables "by hand", using the /etc/oracle/olr.loc file to determine the home of the oracle clusterware.