fixed godb bug for non rac environments, which caused the ORACLE_SID not to be set
The output of the "srvctl status database" is different between oracle rac and oracle restart.
In the latter one, the instance name is not shown, only a message indicating if the database in running or not.

This fix, first checks on the GI type and then parses the instance_name depending on the installation.
The parsing of the srvctl output is now also done via awk (instead of grep and cut), allowing to parse multiple keywords more easily