Modified asmdisks
Changed the hard coded grid home and asm instance name to an invocation of the gogrid script (which path is still hard coded)
This makes it easier when deploying this script on rac environments, where the instance name is different on each node (but the location of the gogrid script probably not).

Changed the sql statement used to list the ongoing operations to use the GV$ views rather than the V$ views.