Remove dependency on oraenv / oratab from godb
In 12.2, the oratab file is no longer maintained by the ora_agent and to avoid having to maintain it manually,
this change will remove the dependency of the godb script on the oratab (and oraenv script).
The ORACLE_HOME value will now be retrieved from the cluster registry and no the oratab.
Also, the LD_LIBRARY_PATH, PATH and ORACLE_BASE variables will be set by the godb script directly and not by calling the oraenv script