Fix Issue 2 - godb fails when the clusterware is a different version than the db software
When querying cluster details for a DB, the used srvctl version must match the DB version, which can differ from the cluster version.
But the DB version \ home is unknown at the start of the script.

To solve this, the clusterware srvctl is used to list all databases together with the ORACLE_HOME (using the "srvctl config database -verbose" command)
and grep for the passed DB unique name.
After that, the further srvctl inquiries are done using the srvctl from the db home for that database.